src/python/requirements.txt
# Python 依赖
grpcio>=1.50.0
grpcio-tools>=1.50.0
opencv-python>=4.8.0
numpy>=1.24.0
